Add files via upload
This firmware adds an event timer functionality. It's not reliable, and may never be, considering the speed restrictions of the hardware, but it's there. You can access the event timer mode by switching to CONFIG mode, holding shift, and pressing buttons 1, 3, and 7. Repeating that will toggle you back to regular quantized mode